forhindre at sjekkboks for beslutter er tilgjengelig for saksbehandler
