+++ app/components/utvidede-beskrivelser/UtvidedeBeskrivelser.tsx
@@ -1,15 +1,9 @@
-import { Checkbox, Detail, Radio, RadioGroup } from "@navikt/ds-react";
-import { useForm } from "@rvf/react-router";
-import { Form } from "react-router";
+import { Detail } from "@navikt/ds-react";
 
 import { UtvidetBeskrivelseInput } from "~/components/utvidede-beskrivelser/UtvidetBeskrivelseInput";
-import { useTypedRouteLoaderData } from "~/hooks/useTypedRouteLoaderData";
-import {
-  hentValideringForGodkjentBrevSkjema,
-  hentValideringForMeldingOmVedtakKildeSkjema,
-} from "~/utils/validering.util";
 
 import { components } from "../../../openapi/melding-om-vedtak-typer";
+import { MeldingOmVedtakKilde } from "../melding-om-vedtak-kilde/MeldingOmVedtakKilde";
 
 type IUtvidetBeskrivelse = components["schemas"]["UtvidetBeskrivelse"];
 
@@ -18,28 +12,6 @@ export function UtvidedeBeskrivelser(props: {
   setUtvidedeBeskrivelser: (utvidedeBeskrivelser: IUtvidetBeskrivelse[]) => void;
   readOnly?: boolean;
 }) {
-  const { oppgave } = useTypedRouteLoaderData("routes/oppgave.$oppgaveId");
-  const meldingOmVedtakKildeSkjema = useForm({
-    method: "post",
-    schema: hentValideringForMeldingOmVedtakKildeSkjema(),
-    defaultValues: {
-      oppgaveId: oppgave.oppgaveId,
-      meldingOmVedtakKilde: oppgave.meldingOmVedtakKilde,
-    },
-  });
-
-  const godkjentBrevSkjema = useForm({
-    method: "post",
-    schema: hentValideringForGodkjentBrevSkjema(),
-    defaultValues: {
-      oppgaveId: oppgave.oppgaveId,
-    },
-  });
-
-  function handleChange() {
-    godkjentBrevSkjema.submit();
-  }
-
   function oppdaterUtvidetBeskrivelse(oppdatertBeskrivelse: IUtvidetBeskrivelse) {
     let oppdatertUtvidedeBeskrivelser = [...props.utvidedeBeskrivelser];
 
@@ -61,45 +33,7 @@ export function UtvidedeBeskrivelser(props: {
 
   return (
     <div className="flex flex-col gap-6">
-      <Form {...meldingOmVedtakKildeSkjema.getFormProps()}>
-        <input name="_action" value="lagre-melding-om-vedtak-kilde" hidden={true} readOnly={true} />
-        <input
-          {...meldingOmVedtakKildeSkjema.field("oppgaveId").getInputProps()}
-          hidden={true}
-          readOnly={true}
-        />
-        <RadioGroup
-          {...meldingOmVedtakKildeSkjema.field("meldingOmVedtakKilde").getInputProps()}
-          size="small"
-          readOnly={props.readOnly}
-          onChange={() => meldingOmVedtakKildeSkjema.submit()}
-          error={meldingOmVedtakKildeSkjema.field("meldingOmVedtakKilde").error()}
-          legend="Send melding om vedtak til bruker"
-        >
-          <Radio value="DP_SAK">Som brev via DP-sak</Radio>
-          <Radio value="GOSYS">Som brev via Gosys</Radio>
-          <Radio value="INGEN">Ikke send melding til bruker</Radio>
-        </RadioGroup>
-      </Form>
-      {oppgave.meldingOmVedtakKilde == "GOSYS" && (
-        <Form {...godkjentBrevSkjema.getFormProps()}>
-          <input name="_action" value="lagre-godkjent-brev-i-gosys" hidden={true} readOnly={true} />
-          <input
-            {...godkjentBrevSkjema.field("oppgaveId").getInputProps()}
-            hidden={true}
-            readOnly={true}
-          />
-          <Checkbox
-            {...godkjentBrevSkjema.field("godkjentBrev").getInputProps()}
-            onChange={handleChange}
-            size="small"
-            defaultChecked={oppgave.kontrollertBrev === "JA"}
-          >
-            Beslutter har godkjent brev i Gosys
-          </Checkbox>
-        </Form>
-      )}
-      <hr className="border-(--a-border-subtle)" />
+      <MeldingOmVedtakKilde readOnly={props.readOnly} />
       {props.utvidedeBeskrivelser.map((utvidetBeskrivelse) => (
         <UtvidetBeskrivelseInput
           key={utvidetBeskrivelse.brevblokkId}
